+++ TombEngine/Scripting/Internal/TEN/View/ViewHandler.cpp
@@ -19,10 +19,12 @@
 #include "Scripting/Internal/TEN/View/ScaleModes.h"
 #include "Scripting/Internal/TEN/View/PostProcessEffects.h"
 #include "Specific/clock.h"
+#include "Specific/Video/Video.h"
 
 using namespace TEN::Effects::Environment;
 using namespace TEN::Scripting::DisplaySprite;
 using namespace TEN::Scripting::View;
+using namespace TEN::Video;
 
 using TEN::Renderer::g_Renderer;
 
@@ -111,6 +113,17 @@ namespace TEN::Scripting::View
 		InitializeSpotCam(seqID);
 	}
 
+	static void PlayVideo(const std::string& fileName, TypeOrNil<bool> background, TypeOrNil<bool> silent, TypeOrNil<bool> loop)
+	{
+		auto mode = ValueOr<bool>(background, false) ? VideoPlaybackMode::Background : VideoPlaybackMode::Exclusive;
+		g_VideoPlayer.Play(fileName, mode, ValueOr<bool>(silent, false), ValueOr<bool>(loop, false));
+	}
+
+	static void StopVideo()
+	{
+		g_VideoPlayer.Stop();
+	}
+
 	static Vec3 GetFlybyPosition(int seqID, float progress, TypeOrNil<bool> loop)
 	{
 		constexpr auto PROGRESS_MAX = 100.0f;
@@ -180,8 +193,8 @@ namespace TEN::Scripting::View
 
 		///Move black cinematic bars in from the top and bottom of the game window.
 		//@function SetCineBars
-		//@tparam float height __(default 30)__ Percentage of the screen to be covered
-		//@tparam float speed __(default 30)__ Coverage percent per second
+		//@tparam float height (default 30). Percentage of the screen to be covered
+		//@tparam float speed (default 30). Coverage percent per second
 		tableView.set_function(ScriptReserved_SetCineBars, &SetCineBars);
 
 		///Set field of view.
@@ -234,6 +247,18 @@ namespace TEN::Scripting::View
 		//@function SetPostProcessTint
 		//@tparam Color tint value to use.
 		tableView.set_function(ScriptReserved_SetPostProcessTint, &SetPostProcessTint);
+
+		/// Play a video file. Should be placed in the `FMV` folder.
+		// @function PlayVideo
+		// @tparam string fileName Video file name.  Can be provided without extension, if type is mp4, mkv or avi.
+		// @tparam[opt] bool background (default: false). Play video in the background mode. In such case, video must be shown using @{View.DisplaySprite}.
+		// @tparam[opt] bool silent (default: false). Play video without sound.
+		// @tparam[opt] bool loop (default: false). Play video in a loop.
+		tableView.set_function(ScriptReserved_PlayVideo, &PlayVideo);
+
+		/// Stop the currently playing video. Only possible if video is playing in the background mode.
+		// @function StopVideo
+		tableView.set_function(ScriptReserved_StopVideo, &StopVideo);
 
 		/// Play a flyby sequence.
 		// @function PlayFlyby